pyproject.toml
[tool.poetry]
name = "flsk"
version = "0.1.0"
description = "Python flask"
authors = ["imausera"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
pyyaml = "^6.0"
pycryptodome = "^3.16.0"
flask-sqlalchemy = "^3.0.2"
cx-oracle = "^8.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
